Model View Presenter Architecture
This commit contains a preliminary write up of
the system architecture for HotdogNotDog. This
adheres to the Model View Presenter design pattern.
The presenter will contain an instance of the
IHotDogView, and will update the view according to
it's hot dog status. The View is not entirely finsished
since I don't exactly know what assets to use. The
presenter will use the results of the IHotDogIdentifier,
which will say if something is a hotdog or not. Link
to lucidchart is https://www.lucidchart.com/invitations/accept/5068a356-a908-4d02-bb30-bb6fea071775