Include UML link in readme
I added a link to a UML that describes our system architecture. I figured that since it is subject to change, I'd rather have a link to it than constantly update an image file. The diagram is hosted on LucidChart, and the current link permissions are view only (since I can't remember if this repo is public or not). Slack me if you want me to add you as a collaborator on the diagram.